Change password hashing to use scrypt
I switched the password hashing over to use libsodium's scrypt
implementation. This involved adding a new migration.

I noticed that currently the password isn't checked. I started looking
into how to implement password checking and it looks like we will
probably need to have a launcher to do so.

The reason the commit touches so many files is because it also cleans up
the files with mixed line endings.