Add automatic database migrations to startup
Database migrations will automatically be run.
Name them according to the format (`#_YYYY-MM-DD-descriptive_name.sql`)
and put in the appropriate directory and it will run on the next server
startup.

Migrations are only run once, in ascending numeric order. The current
database state is tracked in a new database called `Global`.

This simplifies server setup: now the only step is to create the four
databases. The base SQL and migrations will be run automatically from
that point on.

This also simplifies migrations, which will no longer need to
incorporate complex conditional logic in order to be repeatable.