onUse emotes
This code works for all collectors and should work for many more things. The only problem is is that there are emote tables missing. Also, with quest NPCs, the first emote for onUse is to update the quest. Once the quest is updated, It will then give dialogue about the quest. We do not have quest tables in to update it, so you will not see any information regarding quests.