Rename SkillStatus to SkillAdvancementClass
SkillStatus and SkillAdvancementClass are neither the perfect names for what this value represents.

However, the client calls it SkillAdvancementClass. We call it SAC in the database. And GDLE also uses the SAC naming.

I think switching this name helps align ACE with the pdb without adding any additional confusion, meaning, this new name is equally as confusing as the previous, but at last its aligned with the pdb and gdle.