Recipe_Mod Health/Stamina/Mana cleaned up
There were 3 redundant values which were basically flags indicating if it's associated vital was actually used.

However, in the cache.bin, any time a vital was present, it's associated flag was also set to 1. This is basically redundant.

We can simplify this as follows. If the Vital != 0, it gets applied. If the Vital == 0, it doesn't get applied.

I suspect they may have wanted the flag so they could leave a vital value in the db, but just have it disabled. We don't need this.