EnchantmentRegistry improvements
Thread safety added to EnchantmentManager
Thread safety added to EnchantmentRegistry

The locks are required only when we iterate/insert/remove from the collections. EF accesses these collections on a separate thread and they cannot be modified while being iterated over.

Player save removed from EnchantmentManager. Saving a player should be a function of the player, not the enchantment manager. If we want to save a player on death, we can do that from the player.

Some minor cosmetic code improvements here and there.