Collection Modified exception fix
GetPhysicsWorldObjects was returning a linq query. Think of this as a living entity.

As the results were being iterated over, and processed, the processing of a specific result would end up adding a new entity to worldObjects, thus, modifying the backing of the linq query.

By doing a .ToList() on the query, we force the entire query to execute and return all results.