WorldObject Position/Cache cleanup
The positionCache (formerly named Positions) should only be referenced by the Get/Set/Remove functions.

Furthermore, on save, we make sure that the positions in the biota have the latest up to date values from the cache.

We must do this because we don't wrap Positions like we do attributes/skills/vitals.