GameMessagePlayerTeleport was malformed (#1056)
Furthermore, in retail, the sequence sent in GameMessagePlayerTeleport is always the PREVIOUS sequence value, and then the sequence get's incremented AFTER the message is actually secent (and thus, sent in the next UpdatePosition message).

For simplicity, ACE just increments the sequence BEFORE this message is sent. It doesn't seem to affect the client or decal and is cleaner this way.