Fix for vendors selling and/or buying items with Value of 0
Items sold or bought by vendors that report Value of 0 are treated as if they were Value of 1.

This is seen with the vendor in training academy, Oil of Rendering is meant to cost 1p, while its value is 0.

In practice this is rare, and client tends to not allow sale of 0 value items

Co-Authored-By: gmriggs <gmriggs@gmail.com>