ShardDbContextWithFinalize Test
DbContext does not implement a finalizer, so nothing to call Dipose() for us if we fail to do so ourselves (which we are currentlyl...)

This is just a test to see if it helps CE memory usage

This might actually be good for master, but, on windows server core 2016, i haven't noticed the need to explicitly add the finalizer. Even without it, my objects seem to get cleaned up. Could just be my test environment isn't exaggerating the underlying issue though.