Release NetworkSession resources when a session drops
This is a simple patch to help CE mem usage. The proper fix is to funnel NetworkSession work through Session, so that Session can manage it's state directly.

What is happening now, and can/will happen in the future, is that Player references are being held, and thus, their Session and NetworkSession objects aren't being released.

Even after a Session drops, work can still be enqueued onto NetworkSession through ActionChains and other means. This PR will now ignore that work. Currently, in master, that work is received and held indefinitely (until Player/Session are released)

This also fixes the case where NetworkSession work was being enqueued after a session dropped, but before the Player was fully logged out of the world. Now, that work will be ignored.