Make Dat SubPalette and TextureMapChanges properties public
This allows other tools to instantiate these objects and set new values.

This is used by OptimShis Palette decoder and my retail character importer.

I was always on the fence about these dat properties being public. The risk is that a developer could modify a property in ACE. That property would then persist to the cached at object, and would be re-used until ACE was restarted.

This happens because dat objects are cached, and that same instance is returned on subsequent requests.

However, if we are ever to have 3rd party Dat editor tools, these properties will need to be public anyway.