Create partial implementation of HeartbeatManager
This is a prety simple *Manager class that can send periodic requests to another server as a form of heartbeat. This is generally useful and could be used by a server to report uptime, player info, and so on.

Not qutie finished yet here but this works for my testing.