Filter Emotes when IsDormant
This adds a IsDormant check on top of EmoteType.Motion, which already checked for PlayersInRange(). This should not change the result, but instead simply provide a more efficient way to filter this emote. IsDormant is less CPU then calling PlayersInRange, and there should never be PlayersInRange when IsDormant.

EmoteType.Move adds an IsDormant check.

These changes bring current master + 219 content to equal if not better performance than current master + 216 content.