Town Control basic functionality
This is an initial commit of the Town Control functionality that existed on Podtide.

It does not include the item rewards from Podtide, like Ground Anus, A Fuckin Bang, any of the Morph Gems, etc.  It also does not include the code to support Morph Gems - that can be a separate commit.  It does include the reward vendor weenies but they are not placed on landblocks.

This commit is a good starting point if any server wants to use this feature.  For Doctide though, we'll remove the town control currency "A Phial of Innocent Tears" and instead reward PK Trophies; this will be done in a separate commit.  And we'll likely move the town conflict locations to dungeons, and change the rewards and reward vendors quite a bit.