Add ForceLogoffStuckCharacter player command
This command allows a player to force logoff a character that's stuck in game on the same account.  The target has to be on the same account and cannot be the character that issued the command.  Normally its not possible to log in two characters on the same account, but when characters get stuck in game it ends up being possible.