Healing nerf
Heal amount is now limited to the amount of HP that was missing at the start of the heal, and does not include the damage taken during the heal animation.

So for example if you have 200 total hp and 100 remaining hp, you start a heal, then you take an additional 50 dmg during the heal animation.  Previously the heal amount could cover that entire 150 hp you are missing.  Now the max you can heal is the 100 hp you were missing at the start of the heal.

Also the skill modifier for spec healing is reduced from 150% to 130%