Improve ConnectionListener.Listen() to avoid stack overflow
In rare cases, when Listen() continues to throw a SocketException and call back on itself, and throw and call itself and throw... eventual result is stack overflow.

This change simply returns a bool to indicate of the function should be called again, or the Listen() process should be abandoned.

Note, that even currently, the stack is completely reset on a successful Listen()/OnDataReceive(), as each OnDataReceive() is called as a callback, and thus, not tacked onto the calling stack.