Heal difficulty should be based on HP missing at start of heal animation
In a previous commit, we changed healing to only allow you to heal for the amount of HP that was missing at the start of the heal animation.  But the difficulty calculation was still using the amount of HP missing at the end of the animation.  Changed it so difficulty uses the amount of HP missing at the beginning of the heal.