Fix ACE defect that caused a crash
We saw an issue in ACE CalculateDeathItems code (not part of our customizations) where the killer is looked up by GUID based on the KillerID on the corpse, and then its case to a Player object without checking first if the IPlayer is a Player or OfflinePlayer.  Seems like if the killer becomes offline before this code runs, you'd hit this exception and it'd crash the server.  I put in a try/catch as well as a check to see if the iPlayer is a Player before casting it.