Adds use_portal_max_level_requirement server property
This is a minor change adding a server property to optionally disable
the maximum level restriction on portals. Other restrictions are
unaffected by this property.

Server owners can use this property to enable access to certain legacy
content that is otherwise practically inaccessible, such as dungeons and
quests associated with retired starter towns.