Performance improvements to Physics.Common.WeenieObject
Entire set of fields can be marked as readonly, as it never changes outside of the constructor.

Reduces unnecessary reflection operations by pre-fetching them during constructor

Changes WeenieObject to a sealed class to guarantee immutability

Immutability allows a shared DummyObject to be used by PhysicsObjs which do not have a weenie.